sing-email-with-certificates/requirements.txt
python-smail==0.9.0